Time: 2 ms (56.28%), Space: 7.1 MB (52.25%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -6,7 +6,7 @@ public:
 
         while(start <= end)
         {
-            long long mid = start + (end - start) / 2;
+            long mid = start + (end - start) / 2;
 
             if(mid * mid == num)
             return true;
